+++ project/src/reservation/components/AccomodationInfo.js
